+++ src/OpenLoco/src/ScenarioManager.cpp
@@ -369,7 +369,7 @@ namespace OpenLoco::ScenarioManager
     }
 
     // 0x0044452F
-    void loadIndex(uint8_t al)
+    void loadIndex(const bool forceReload)
     {
         const auto oldFlags = getScreenFlags();
         setAllScreenFlags(ScreenFlags::title | oldFlags);
@@ -380,18 +380,11 @@ namespace OpenLoco::ScenarioManager
         }
 
         const auto currentState = getCurrentScenarioFolderState();
-
-        if (!tryLoadIndex(currentState))
+        if (forceReload || !tryLoadIndex(currentState))
         {
             createIndex(currentState);
         }
-        else
-        {
-            if (al != 0)
-            {
-                createIndex(currentState);
-            }
-        }
+
         setAllScreenFlags(oldFlags);
     }